planning: fix compile warning.

+++ modules/planning/math/curve1d/quartic_polynomial_curve1d.cc
@@ -135,7 +135,7 @@ QuarticPolynomialCurve1d& QuarticPolynomialCurve1d::IntegratedFromCubicCurve(
   param_ = other.ParamLength();
   coef_[0] = init_value;
   for (std::size_t i = 0; i < 4; ++i) {
-    coef_[i + 1] = other.Coef(i) / (i + 1);
+    coef_[i + 1] = other.Coef(i) / (static_cast<double>(i) + 1);
   }
   return *this;
 }
@@ -145,7 +145,7 @@ QuarticPolynomialCurve1d& QuarticPolynomialCurve1d::DerivedFromQuinticCurve(
   CHECK_EQ(other.Order(), 5);
   param_ = other.ParamLength();
   for (std::size_t i = 1; i < 6; ++i) {
-    coef_[i - 1] = other.Coef(i) * i;
+    coef_[i - 1] = other.Coef(i) * static_cast<double>(i);
   }
   return *this;
 }
